requirements.txt
aiohttp
aiofiles
asyncpg
Pillow
pytest
beautifulsoup4
lxml
Flask
psycopg2-binary